+++ pkg/ingester/ingester.go
@@ -262,7 +262,7 @@ func (i *Ingester) startFlushLoops() {
 // Compared to the 'New' method:
 //   * Always replays the WAL.
 //   * Does not start the lifecycler.
-func NewForFlusher(cfg Config, chunkStore ChunkStore, registerer prometheus.Registerer) (*Ingester, error) {
+func NewForFlusher(cfg Config, chunkStore ChunkStore, limits *validation.Overrides, registerer prometheus.Registerer) (*Ingester, error) {
 	if cfg.BlocksStorageEnabled {
 		return NewV2ForFlusher(cfg, registerer)
 	}
@@ -273,9 +273,10 @@ func NewForFlusher(cfg Config, chunkStore ChunkStore, registerer prometheus.Regi
 		chunkStore:  chunkStore,
 		flushQueues: make([]*util.PriorityQueue, cfg.ConcurrentFlushes),
 		wal:         &noopWAL{},
+		limits:      limits,
 	}
 
-	i.BasicService = services.NewBasicService(i.startingForFlusher, i.loop, i.stopping)
+	i.BasicService = services.NewBasicService(i.startingForFlusher, i.loopForFlusher, i.stopping)
 	return i, nil
 }
 
@@ -295,6 +296,18 @@ func (i *Ingester) startingForFlusher(ctx context.Context) error {
 
 	i.startFlushLoops()
 	return nil
+}
+
+func (i *Ingester) loopForFlusher(ctx context.Context) error {
+	for {
+		select {
+		case <-ctx.Done():
+			return nil
+
+		case err := <-i.subservicesWatcher.Chan():
+			return errors.Wrap(err, "ingester subservice failed")
+		}
+	}
 }
 
 func (i *Ingester) loop(ctx context.Context) error {
@@ -710,6 +723,7 @@ func (i *Ingester) QueryStream(req *client.QueryRequest, stream client.Ingester_
 	}
 
 	numSeries, numChunks := 0, 0
+	reuseWireChunks := [queryStreamBatchSize][]client.Chunk{}
 	batch := make([]client.TimeSeriesChunk, 0, queryStreamBatchSize)
 	// We'd really like to have series in label order, not FP order, so we
 	// can iteratively merge them with entries coming from the chunk store.  But
@@ -728,10 +742,12 @@ func (i *Ingester) QueryStream(req *client.QueryRequest, stream client.Ingester_
 		}
 
 		numSeries++
-		wireChunks, err := toWireChunks(chunks, nil)
+		reusePos := len(batch)
+		wireChunks, err := toWireChunks(chunks, reuseWireChunks[reusePos])
 		if err != nil {
 			return err
 		}
+		reuseWireChunks[reusePos] = wireChunks
 
 		numChunks += len(wireChunks)
 		batch = append(batch, client.TimeSeriesChunk{